// src/app/module/client/Vos/PODetail.ts
import { ProjectDetail } from "./projectDetailmodel";

export class PODetail {
      id;

      name: string;

      number: number;

      amount: number;

      attachmentName: string;

      attachmentPath: string;

      project: ProjectDetail;

      createdBy;

      updatedBy;
}